+++ src/pyjapi/__init__.py
@@ -1,3 +1,3 @@
-# JAPIClient needs to be imported first
-from .cli import __version__, cli  # noqa
+from ._types import *  # noqa
+from .cli import __version__  # noqa
 from .lib import JAPIClient  # noqa